Move community module into its own subfolder so we can add other possible projects too
