+++ CommunityModule/Static Files/English.dat
@@ -1,0 +1,2 @@
+Name Community
+Description Collection of community-contributed code allowing some clientside modding.
