--- English.dat
@@ -1,2 +1,0 @@
-Name Community
-Description Collection of community-contributed code allowing some clientside modding.
