--- lib/.gitignore
@@ -1,3 +1,0 @@
-Assembly-CSharp.dll
-Assembly-CSharp-firstpass.dll
-UnityEngine.dll
