+++ unturned-lib/required-files.txt
@@ -1,0 +1,4 @@
+You need to copy the following files from "Unturned/Unturned_Data/Managed" to here:
+Assembly-CSharp.dll
+Assembly-CSharp-firstpass.dll
+UnityEngine.dll
